+++ pub/unrouter/test/exports_test.dart
@@ -7,5 +7,6 @@ void main() {
     expect(unrouter.RedirectLoopPolicy.values, isNotEmpty);
     expect(unrouter.UnrouterResolutionState.values, isNotEmpty);
     expect(unrouter.UnrouterStateTimelineEntry, isNotNull);
+    expect(unrouter.UnrouterController, isNotNull);
   });
 }